pyproject.toml
[project]
name = "argus"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bytes>=0.0.1.dev4",
    "can-isotp>=2.0.7",
    "pyserial>=3.5",
    "python-can>=4.6.1",
]

[tool.uv]
package = true

[project.scripts]
ui = "argus.ui:main"

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "DEBUG"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
log_file = "logs/pytest_run.log"
log_file_level = "INFO"
log_file_format = "%(asctime)s - %(name)s %(levelname)s %(message)s"
log_file_date_format = "%Y-%m-%d %H:%M:%S"

[dependency-groups]
dev = ["pytest>=8.4.2"]
